Reduce configalchemy to one file
