+++ configalchemy.py
@@ -1,9 +1,15 @@
+#!/usr/bin/env python
+# coding: utf8
+from __future__ import unicode_literals
+
 import sys
+import argparse
 import itertools
 
 from configparser import (BasicInterpolation, DuplicateOptionError,
                           DuplicateSectionError, SectionProxy,
                           MissingSectionHeaderError)
+
 from clldutils.inifile import INI as ClldIni
 
 
@@ -214,3 +220,39 @@ class INI (ClldIni):
                     val = '\n'.join(val).rstrip()
                 options[name] = val
 
+
+def main(*args):
+    """Execute CLI functionality
+
+    Parse command line arguments to find config files and distill a an
+    aggregated, interpolated configuration from them.
+
+    """
+    # Parse command line arguments
+    parser = argparse.ArgumentParser(
+        description="Combine multiple configuration files into one.")
+    parser.add_argument(
+        "config",
+        type=argparse.FileType("r"),
+        help="ConfigParser-style configuration file(s)",
+        default=None,
+        nargs="+")
+    parser.add_argument(
+        "-o", "--output",
+        type=argparse.FileType("w"),
+        help="Output filename, for example `-o gold.ini`",
+        default=sys.stdout)
+    args = parser.parse_args(args or None)
+    
+    config = INI(interpolation=BasicReadInterpolation())
+    for file in args.config:
+        config.read_file(file)
+
+    # We would like to use clldutils.INI.write directly, but it does
+    # not support file objects.
+    args.output.write(config.write_string())
+    
+    sys.exit(0)
+
+if __name__ == "__main__":
+    main(sys.argv)
